+++ brewtils/models.py
@@ -1510,6 +1510,7 @@ class Garden(BaseModel):
         metadata=None,
         default_user=None,
         shared_users=None,
+        version="0.0.0",
     ):
         self.id = id
         self.name = name
@@ -1530,16 +1531,19 @@ class Garden(BaseModel):
         self.default_user = default_user
         self.shared_users = shared_users
 
+        self.version = version
+
     def __str__(self):
         return "%s" % self.name
 
     def __repr__(self):
         return (
-            "<Garden: garden_name=%s, status=%s, parent=%s, has_parent=%s, "
+            "<Garden: garden_name=%s, status=%s, version=%s, parent=%s, has_parent=%s, "
             "connection_type=%s, receiving_connections=%s, publishing_connections=%s>"
             % (
                 self.name,
                 self.status,
+                self.version,
                 self.parent,
                 self.has_parent,
                 self.connection_type,